Move components in Table to components folder
