+++ src/components/Table/DrillDown/components/DrillDownTable.tsx
@@ -2,7 +2,7 @@ import { Dictionary } from '@onaio/utils';
 import React, { useState } from 'react';
 import { Cell, Column } from 'react-table';
 import Ripple from '../../../page/Loading';
-import { DropDownCell, DropDownCellProps } from '../helpers/DefaultDropDownCell';
+import { DropDownCell, DropDownCellProps } from './DefaultDropDownCell';
 import { defaultTableProps, Table, TableJSXProps } from './TableJSX';
 
 /** describes props for the DrillDownTable component */
